Move `BitPosition` into its own file.
This will allow the GC info encoder and the JIT to use this
function without pulling in the entire utilcode header.
